Add consolidate tracks updater for processing batches
